// ui/src/proto/appmodel.ts
export interface ApplicationModelEvent {
  clusterName: string;
  nodeName: string;
  time: Date | undefined;
  applicationModel: ApplicationModel | undefined;
}

export interface ApplicationModel {
  namespaces: ApplicationNamespace[];
  host: ApplicationHost | undefined;
}

export interface ApplicationHost {
  processes: ApplicationProcess[];
}

export interface ApplicationNamespace {
  name: string;
  workloads: ApplicationWorkload[];
}
// ...
export const ApplicationModelEvent: MessageFns<ApplicationModelEvent> = {
  encode(message: ApplicationModelEvent, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.clusterName !== "") {
      writer.uint32(10).string(message.clusterName);
    }
    if (message.nodeName !== "") {
      writer.uint32(18).string(message.nodeName);
    }
    if (message.time !== undefined) {
      Timestamp.encode(toTimestamp(message.time), writer.uint32(26).fork()).join();
    }
    if (message.applicationModel !== undefined) {
      ApplicationModel.encode(message.applicationModel, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationModelEvent {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationModelEvent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.clusterName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.nodeName = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.time = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.applicationModel = ApplicationModel.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationModelEvent {
    return {
      clusterName: isSet(object.clusterName) ? globalThis.String(object.clusterName) : "",
      nodeName: isSet(object.nodeName) ? globalThis.String(object.nodeName) : "",
      time: isSet(object.time) ? fromJsonTimestamp(object.time) : undefined,
      applicationModel: isSet(object.applicationModel) ? ApplicationModel.fromJSON(object.applicationModel) : undefined,
    };
  },

  toJSON(message: ApplicationModelEvent): unknown {
    const obj: any = {};
    if (message.clusterName !== "") {
      obj.clusterName = message.clusterName;
    }
    if (message.nodeName !== "") {
      obj.nodeName = message.nodeName;
    }
    if (message.time !== undefined) {
      obj.time = message.time.toISOString();
    }
    if (message.applicationModel !== undefined) {
      obj.applicationModel = ApplicationModel.toJSON(message.applicationModel);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationModelEvent>, I>>(base?: I): ApplicationModelEvent {
    return ApplicationModelEvent.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationModelEvent>, I>>(object: I): ApplicationModelEvent {
    const message = createBaseApplicationModelEvent();
    message.clusterName = object.clusterName ?? "";
    message.nodeName = object.nodeName ?? "";
    message.time = object.time ?? undefined;
    message.applicationModel = (object.applicationModel !== undefined && object.applicationModel !== null)
      ? ApplicationModel.fromPartial(object.applicationModel)
      : undefined;
    return message;
  },
};

function createBaseApplicationModel(): ApplicationModel {
  return { namespaces: [], host: undefined };
}

export const ApplicationModel: MessageFns<ApplicationModel> = {
  encode(message: ApplicationModel, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.namespaces) {
      ApplicationNamespace.encode(v!, writer.uint32(10).fork()).join();
    }
    if (message.host !== undefined) {
      ApplicationHost.encode(message.host, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationModel {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationModel();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.namespaces.push(ApplicationNamespace.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.host = ApplicationHost.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationModel {
    return {
      namespaces: globalThis.Array.isArray(object?.namespaces)
        ? object.namespaces.map((e: any) => ApplicationNamespace.fromJSON(e))
        : [],
      host: isSet(object.host) ? ApplicationHost.fromJSON(object.host) : undefined,
    };
  },

  toJSON(message: ApplicationModel): unknown {
    const obj: any = {};
    if (message.namespaces?.length) {
      obj.namespaces = message.namespaces.map((e) => ApplicationNamespace.toJSON(e));
    }
    if (message.host !== undefined) {
      obj.host = ApplicationHost.toJSON(message.host);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationModel>, I>>(base?: I): ApplicationModel {
    return ApplicationModel.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationModel>, I>>(object: I): ApplicationModel {
    const message = createBaseApplicationModel();
    message.namespaces = object.namespaces?.map((e) => ApplicationNamespace.fromPartial(e)) || [];
    message.host = (object.host !== undefined && object.host !== null)
      ? ApplicationHost.fromPartial(object.host)
      : undefined;
    return message;
  },
};

function createBaseApplicationHost(): ApplicationHost {
  return { processes: [] };
}

export const ApplicationHost: MessageFns<ApplicationHost> = {
  encode(message: ApplicationHost, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.processes) {
      ApplicationProcess.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationHost {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationHost();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.processes.push(ApplicationProcess.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationHost {
    return {
      processes: globalThis.Array.isArray(object?.processes)
        ? object.processes.map((e: any) => ApplicationProcess.fromJSON(e))
        : [],
    };
  },

  toJSON(message: ApplicationHost): unknown {
    const obj: any = {};
    if (message.processes?.length) {
      obj.processes = message.processes.map((e) => ApplicationProcess.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationHost>, I>>(base?: I): ApplicationHost {
    return ApplicationHost.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationHost>, I>>(object: I): ApplicationHost {
    const message = createBaseApplicationHost();
    message.processes = object.processes?.map((e) => ApplicationProcess.fromPartial(e)) || [];
    return message;
  },
};

function createBaseApplicationNamespace(): ApplicationNamespace {
  return { name: "", workloads: [] };
}

export const ApplicationNamespace: MessageFns<ApplicationNamespace> = {
  encode(message: ApplicationNamespace, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    for (const v of message.workloads) {
      ApplicationWorkload.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ApplicationNamespace {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseApplicationNamespace();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.workloads.push(ApplicationWorkload.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ApplicationNamespace {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      workloads: globalThis.Array.isArray(object?.workloads)
        ? object.workloads.map((e: any) => ApplicationWorkload.fromJSON(e))
        : [],
    };
  },

  toJSON(message: ApplicationNamespace): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.workloads?.length) {
      obj.workloads = message.workloads.map((e) => ApplicationWorkload.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ApplicationNamespace>, I>>(base?: I): ApplicationNamespace {
    return ApplicationNamespace.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ApplicationNamespace>, I>>(object: I): ApplicationNamespace {
    const message = createBaseApplicationNamespace();
    message.name = object.name ?? "";
    message.workloads = object.workloads?.map((e) => ApplicationWorkload.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}